Refactor _init_ui and stlt into focused helper functions
Item 1c: Extract _init_ui (1469 lines) into 12 tab-builder functions
in gwexpy/gui/ui/_tab_builders.py. The _init_ui method is now ~40 lines
that orchestrates tab construction via imported helpers.

Item 1f: Refactor stlt method (470 lines) to use the existing extracted
helpers (_resolve_stlt_params, _chunk_data, _prepare_sigma,
_configure_frequency_domain, _prepare_decay, _stlt_main_loop,
_apply_stlt_scaling, _create_stlt_container). The method is now ~76 lines.
