+++ gwexpy/timeseries/_spectral_special.py
@@ -22,14 +22,6 @@ if TYPE_CHECKING:
 # ---------------------------------------------------------------------------
 
 
-def _stlt_legacy_fallback(self, **kwargs):
-    """Handle legacy STLT mode by delegating to ``_stlt_legacy``."""
-    # Extract the positional-style args that stlt() forwards
-    stride = kwargs.pop("stride", None)
-    window = kwargs.pop("window", None)
-    return self._stlt_legacy(stride, window, **kwargs)
-
-
 def _resolve_stlt_params(stride, window, fftlength, overlap, self):
     """Resolve STLT time parameters and return resolved values.
 
@@ -752,99 +744,19 @@ class TimeSeriesSpectralSpecialMixin(TimeSeriesAttrs):
         if legacy:
             return self._stlt_legacy(stride, window, **kwargs)
 
-        try:
-            import scipy.signal  # noqa: F401 - availability check
-        except ImportError:
-            raise ImportError("scipy is required for stlt.")
+        import scipy.signal
 
-        # --- 1. Resolve Time Parameters ---
-        if self.dt is None:
-            raise ValueError("TimeSeries must have a valid dt.")
-        dt_s = self.dt.to("s").value
-        1.0 / dt_s
+        # --- 1. Resolve time parameters ---
+        params = _resolve_stlt_params(stride, window, fftlength, overlap, self)
+        dt_s = params["dt_s"]
+        nperseg = params["nperseg"]
+        step = params["step"]
+        win_dur = params["win_dur"]
+        window_func = params["window_func"]
 
-        # Helper to ensure quantity is in seconds
-        def _to_sec(val, name):
-            if val is None:
-                return None
-            if isinstance(val, (int, float, np.number)):
-                return float(val)
-            q = u.Quantity(val)
-            if q.unit == u.dimensionless_unscaled:
-                # Assume seconds if dimensionless (e.g. from float) -> wait, Quantity(0.5) is dimless
-                # Quantity(0.5, 's') is time.
-                # If user passed Number, we handled above.
-                # If user passed Quantity(0.5), treating as seconds is ambiguous but likely intended if not specifed.
-                # But safer to assign unit "s" if original was just number.
-                return q.value
-            try:
-                return q.to("s").value
-            except u.UnitConversionError:
-                raise ValueError(f"{name} must be a time quantity (e.g. seconds).")
-
-        # Resolve fftlength (duration) / window (function vs duration)
-        window_func = "hann"
-        win_dur = None
-
-        if fftlength is not None:
-            win_dur = _to_sec(fftlength, "fftlength")
-            if window is not None and not isinstance(window, (u.Quantity, float, int)):
-                window_func = window
-        elif window is not None:
-            is_dur = False
-            if isinstance(window, (int, float, np.number)) or isinstance(
-                window, u.Quantity
-            ):
-                is_dur = True
-            elif isinstance(window, str) and any(c.isdigit() for c in window):
-                # Heuristic: if string contains digits, treat as quantity string "4s"
-                is_dur = True
-
-            if is_dur:
-                win_dur = _to_sec(window, "window")
-            else:
-                window_func = window
-
-        if win_dur is None:
-            raise ValueError("Must specify fftlength (or window duration).")
-
-        nperseg = int(np.round(win_dur / dt_s))
-        if nperseg < 1:
-            raise ValueError("Window duration too short.")
-
-        # Resolve Overlap/Stride
-        if overlap is not None:
-            ov_dur = _to_sec(overlap, "overlap")
-            noverlap = int(np.round(ov_dur / dt_s))
-            step = nperseg - noverlap
-            str_dur = step * dt_s
-        elif stride is not None:
-            str_dur = _to_sec(stride, "stride")
-            step = int(np.round(str_dur / dt_s))
-
-            # Re-calculate str_dur from integer step to match actual execution?
-            # Or keep user's request? Standard is usually to snap to samples.
-            # step = max(1, step) # ensure at least 1 sample? Validated below.
-
-            noverlap = nperseg - step
-            ov_dur = noverlap * dt_s
-        else:
-            step = nperseg // 2
-            noverlap = nperseg - step
-            str_dur = step * dt_s
-            ov_dur = noverlap * dt_s
-
-        if step < 1:
-            raise ValueError("Stride must be positive (overlap < window).")
-
-        # --- 2. Data Chunking & Pre-checks ---
-        from numpy.lib.stride_tricks import sliding_window_view
-
+        # --- 2. Data chunking ---
         data_arr = self.value
-        if len(data_arr) < nperseg:
-            raise ValueError("Data shorter than window length.")
-
-        chunks = sliding_window_view(data_arr, window_shape=nperseg, axis=0)[::step]
+        chunks = _chunk_data(data_arr, nperseg, step)
         n_chunks = chunks.shape[0]
 
         t0_val = self.t0.to("s").value
@@ -852,7 +764,7 @@ class TimeSeriesSpectralSpecialMixin(TimeSeriesAttrs):
         t_centers = t_starts + (win_dur / 2.0)
         times_q = u.Quantity(t_centers, "s")
 
-        # --- 3. Window & Sigma Prep ---
+        # --- 3. Window function ---
         if isinstance(window_func, (str, tuple)):
             win_base = scipy.signal.get_window(window_func, nperseg)
         else:
@@ -860,300 +772,52 @@ class TimeSeriesSpectralSpecialMixin(TimeSeriesAttrs):
             if len(win_base) != nperseg:
                 raise ValueError("Window function length mismatch.")
 
-        # Handle Sigmas
-        if np.isscalar(sigmas) or (isinstance(sigmas, u.Quantity) and sigmas.isscalar):
-            sigmas_arr = [sigmas]
-        else:
-            sigmas_arr = list(sigmas)
+        # --- 4. Parse sigmas ---
+        sigmas_vals = _prepare_sigma(sigmas, None)
 
-        sigmas_vals_list: list[float] = []
-        for s in sigmas_arr:
-            if isinstance(s, u.Quantity):
-                sigmas_vals_list.append(float(s.to("1/s").value))
-            else:
-                sigmas_vals_list.append(float(s))  # type: ignore[arg-type]
-        sigmas_vals = np.asarray(sigmas_vals_list, dtype=float)
-
-        # Handle One-sided
+        # --- 5. Resolve onesided ---
         is_complex_input = np.iscomplexobj(data_arr)
         if onesided is None:
             onesided = not is_complex_input
-
         if onesided and is_complex_input:
             raise ValueError("Cannot perform one-sided FFT on complex data.")
 
-        # Precompute t_rel
-        if time_ref == "start":
-            t_rel = np.arange(nperseg) * dt_s
-        elif time_ref == "center":
-            # centered around 0
-            t_rel = (np.arange(nperseg) - (nperseg - 1) / 2.0) * dt_s
-        else:
-            raise ValueError(f"Unknown time_ref: {time_ref}. Use 'start' or 'center'.")
+        # --- 6. Decay matrix and stability check ---
+        decay_info = _prepare_decay(sigmas_vals, time_ref, dt_s, nperseg)
+        effective_windows = win_base[None, :] * decay_info["decay_matrix"]
 
-        # Stability Guardrails
-        # Max exponent argument
-        min_sigma = sigmas_vals.min()
-        max_sigma = sigmas_vals.max()
-        max_t = t_rel.max()
-        min_t = t_rel.min()  # negative if centered
-
-        # We care about -sigma * t
-        # Case 1: sigma positive, t negative (if centered) -> -sigma*t positive (growth)
-        # Case 2: sigma negative, t positive -> -sigma*t positive (growth)
-
-        max_exponent = max(
-            -min_sigma * min_t,
-            -min_sigma * max_t,
-            -max_sigma * min_t,
-            -max_sigma * max_t,
-        )
-
-        if max_exponent > 700:  # approx exp(709) is max double
-            raise ValueError(
-                f"Configuration leads to overflow: max exponent ~{max_exponent:.1f} > 700. "
-                f"Try reducing sigma magnitude or using time_ref='center'."
-            )
-
-        # Prepare outputs
-        fs_val = 1.0 / dt_s
-        nyquist = fs_val / 2.0
-        df = fs_val / nperseg
-        tol = df * 1e-9
-
-        def _coerce_frequencies(freqs):
-            if isinstance(freqs, u.Quantity):
-                freqs_val = np.asarray(freqs.to("Hz").value, dtype=float)
-                freqs_q = freqs.to("Hz")
-            else:
-                freqs_val = np.asarray(freqs, dtype=float)
-                freqs_q = u.Quantity(freqs_val, "Hz")
-            if freqs_val.ndim != 1:
-                raise ValueError("frequencies must be a 1D array.")
-            if freqs_val.size == 0:
-                raise ValueError("frequencies must be non-empty.")
-            return freqs_val, freqs_q
-
-        def _validate_frequencies(freqs_val):
-            if onesided:
-                if np.any(freqs_val < -tol) or np.any(freqs_val > nyquist + tol):
-                    raise ValueError(
-                        "frequencies must be within [0, Nyquist] for onesided STLT."
-                    )
-            else:
-                if np.any(freqs_val < -nyquist - tol) or np.any(
-                    freqs_val > nyquist + tol
-                ):
-                    raise ValueError(
-                        "frequencies must be within [-Nyquist, Nyquist] for two-sided STLT."
-                    )
-
-        def _fft_bin_indices(freqs_val):
-            if onesided:
-                k = np.rint(freqs_val / df).astype(int)
-                if np.any(k < 0) or np.any(k > nperseg // 2):
-                    return None
-                if np.max(np.abs(freqs_val - k * df)) > tol:
-                    return None
-                return k
-
-            k = np.rint(freqs_val / df).astype(int)
-            k_mod = k % nperseg
-            grid_freq = np.where(
-                k_mod <= nperseg // 2, k_mod * df, (k_mod - nperseg) * df
-            )
-            if np.max(np.abs(freqs_val - grid_freq)) > tol:
-                return None
-            return k_mod
-
-        freq_mode = "fft"
-        fft_indices = None
-        use_zoom = False
-        zoom_info = None
-        reverse_zoom = False
-        custom_method = None
-        cached_phase = None
-        ZOOM_MIN_M = 256
-        PHASE_CACHE_MAX_M = 256
-
-        if frequencies is None:
-            if onesided:
-                freqs_val = np.fft.rfftfreq(nperseg, d=dt_s)
-            else:
-                freqs_val = np.fft.fftfreq(nperseg, d=dt_s)
-            freqs_q = u.Quantity(freqs_val, "Hz")
-        else:
-            freqs_val, freqs_q = _coerce_frequencies(frequencies)
-            _validate_frequencies(freqs_val)
-
-            # Snap tiny values to 0 to reduce numeric drift
-            freqs_val = np.where(np.abs(freqs_val) < tol, 0.0, freqs_val)
-            freqs_q = u.Quantity(freqs_val, "Hz")
-
-            fft_indices = _fft_bin_indices(freqs_val)
-            if fft_indices is not None:
-                freq_mode = "fft_subset"
-            else:
-                freq_mode = "custom"
-
-                # Check for uniform grid to enable zoom FFT
-                if freqs_val.size >= 2:
-                    diffs = np.diff(freqs_val)
-                    diffs_close = np.allclose(diffs, diffs[0], rtol=0.0, atol=tol * 10)
-                    if diffs_close and not np.isclose(diffs[0], 0.0, atol=tol):
-                        step = diffs[0]
-                        reverse_zoom = step < 0
-                        fmin = freqs_val[-1] if reverse_zoom else freqs_val[0]
-                        fmax = freqs_val[0] if reverse_zoom else freqs_val[-1]
-                        span = fmax - fmin
-                        m = freqs_val.size
-                        step_abs = abs(step)
-                        endpoint = abs(span - step_abs * (m - 1)) <= tol * 10
-                        if endpoint or abs(span - step_abs * m) <= tol * 10:
-                            if m > ZOOM_MIN_M:
-                                use_zoom = True
-                                zoom_info = (fmin, fmax, m, endpoint)
-                                custom_method = "zoom"
-                            else:
-                                custom_method = "dft"
-                        else:
-                            custom_method = "dft"
-                    else:
-                        custom_method = "dft"
-                else:
-                    custom_method = "dft"
-
-                # Precompute phase matrix for small frequency sets (DFT path)
-                if custom_method == "dft" and freqs_val.size <= PHASE_CACHE_MAX_M:
-                    cached_phase = np.exp(
-                        (-2j * np.pi) * (freqs_val[:, None] * t_rel[None, :])
-                    )
-
-        n_freqs = len(freqs_val)
+        # --- 7. Frequency domain configuration ---
+        freq_config = _configure_frequency_domain(frequencies, onesided, dt_s, nperseg)
+        freqs_q = freq_config["freqs_q"]
+        n_freqs = len(freq_config["freqs_val"])
         n_sigmas = len(sigmas_vals)
 
-        dtype = np.result_type(chunks.dtype, np.complex64)
-        out_cube = np.zeros((n_chunks, n_sigmas, n_freqs), dtype=dtype)
+        # --- 8. Main computation loop ---
+        chunk_info = {
+            "chunks": chunks,
+            "effective_windows": effective_windows,
+            "n_chunks": n_chunks,
+            "n_sigmas": n_sigmas,
+            "n_freqs": n_freqs,
+            "nperseg": nperseg,
+            "dt_s": dt_s,
+            "onesided": onesided,
+        }
+        out_cube = _stlt_main_loop(chunk_info, decay_info, freq_config)
 
-        # --- 4. Computation Loop with Batching ---
-        # Batch over chunks to control memory usage
-        # weighted_chunks for 1 sigma = n_chunks * nperseg * 16 bytes (complex128)
-        # We want to keep weighted_chunks < ~100MB?
-        # 100MB / 16 bytes ~ 6e6 elements.
-        # If nperseg=1000, batch_size=6000 chunks.
+        # --- 9. Scaling ---
+        _apply_stlt_scaling(out_cube, scaling, dt_s)
 
-        BATCH_ELEMENTS = 5_000_000
-        # Adjust batch size to account for expansion by n_sigmas
-        # Each element in the batch loop will become (batch_size, n_sigmas, nperseg) complex128
-        # So we divide BATCH_ELEMENTS by (nperseg * n_sigmas)
-        elements_per_chunk = nperseg * n_sigmas
-        chunk_batch_size = max(1, BATCH_ELEMENTS // elements_per_chunk)
-
-        # Precompute effective windows for all sigmas: (n_sigmas, nperseg)
-        # decay = exp(-sigma * t)
-        # shape: (S, N)
-        decay_matrix = np.exp(-sigmas_vals[:, None] * t_rel[None, :])
-        effective_windows = win_base[None, :] * decay_matrix
-
-        for i_chunk in range(0, n_chunks, chunk_batch_size):
-            end_chunk = min(i_chunk + chunk_batch_size, n_chunks)
-            # Get batch of chunks: (batch_size, nperseg)
-            batch_chunks = chunks[i_chunk:end_chunk]  # View
-
-            # Vectorized Window Application
-            # batch_chunks: (B, N)
-            # effective_windows: (S, N)
-            # Broadcasting: (B, 1, N) * (1, S, N) -> (B, S, N)
-            weighted_batch = batch_chunks[:, None, :] * effective_windows[None, :, :]
-
-            if freq_mode == "fft":
-                fft_func = np.fft.rfft if onesided else np.fft.fft
-                # FFT along the last axis (time/window axis)
-                # Result shape: (B, S, n_freqs)
-                spec = fft_func(weighted_batch, axis=-1)
-            elif freq_mode == "fft_subset":
-                fft_func = np.fft.rfft if onesided else np.fft.fft
-                spec_full = fft_func(weighted_batch, axis=-1)
-                spec = spec_full[:, :, fft_indices]
-            elif use_zoom and zoom_info is not None:
-                fmin, fmax, m, endpoint = zoom_info
-                spec = scipy.signal.zoom_fft(
-                    weighted_batch,
-                    (fmin, fmax),
-                    m=m,
-                    fs=fs_val,
-                    endpoint=endpoint,
-                    axis=-1,
-                )
-                if reverse_zoom:
-                    spec = spec[:, :, ::-1]
-            else:
-                # Arbitrary frequency evaluation (direct DFT)
-                if cached_phase is not None:
-                    # (B, S, N) x (F, N) -> (B, S, F)
-                    spec = np.tensordot(weighted_batch, cached_phase, axes=([-1], [1]))
-                else:
-                    freq_chunk_size = max(16, min(256, int(2_000_000 // nperseg) or 16))
-                    spec = np.zeros(
-                        (weighted_batch.shape[0], weighted_batch.shape[1], n_freqs),
-                        dtype=np.result_type(weighted_batch.dtype, np.complex64),
-                    )
-                    for i_freq in range(0, n_freqs, freq_chunk_size):
-                        f_end = min(i_freq + freq_chunk_size, n_freqs)
-                        f_chunk = freqs_val[i_freq:f_end]
-                        phase = np.exp(
-                            (-2j * np.pi) * (f_chunk[:, None] * t_rel[None, :])
-                        )
-                        # (B, S, N) x (F, N) -> (B, S, F)
-                        spec[:, :, i_freq:f_end] = np.tensordot(
-                            weighted_batch, phase, axes=([-1], [1])
-                        )
-
-            # Store
-            out_cube[i_chunk:end_chunk, :, :] = spec
-
-        # --- 5. Scaling ---
-        if scaling == "dt":
-            out_cube *= dt_s
-        elif scaling == "none":
-            pass
-        else:
-            raise ValueError(f"Unknown scaling: {scaling}")
-
-        # --- 6. Container ---
-        from gwexpy.types.array3d import Array3D
-        from gwexpy.types.time_plane_transform import LaplaceGram
-
-        # Axis 1 is Sigma
-        axis_sigma = u.Quantity(sigmas_vals, "1/s")
-        # Axis 2 is Frequency
-        axis_freq = freqs_q
-
-        # Unit
-        res_unit = self.unit
-        if scaling == "dt":
-            res_unit = res_unit * u.s
-
-        # Construct Array3D explicitly
-        arr3d = Array3D(
-            out_cube,
-            unit=res_unit,
-            axis0=times_q,
-            axis1=axis_sigma,
-            axis2=axis_freq,
-            axis_names=["time", "sigma", "frequency"],
-        )
-
-        return LaplaceGram(
-            arr3d,
-            kind="stlt",
-            meta={
-                "window": win_dur,
-                "stride": str_dur,
-                "overlap": ov_dur,
-                "source": self.name,
-            },
-        )
+        # --- 10. Build output container ---
+        metadata = {
+            "scaling": scaling,
+            "win_dur": win_dur,
+            "str_dur": params["str_dur"],
+            "ov_dur": params["ov_dur"],
+            "source_name": self.name,
+            "source_unit": self.unit,
+        }
+        return _create_stlt_container(out_cube, times_q, freqs_q, sigmas_vals, metadata)
 
     def _stlt_legacy(self, stride: Any, window: Any, **kwargs: Any) -> Any:
         # Copied from original implementation
